Big changes.
In order to reduce code duplication, I moved the "core" of all the
functions into a new file, and then the main functions are all
implemented in terms of those core functions. This was already done for
remquo and floor/ceil/trunc. In those cases, I moved the core function
to the new header file, and then moved the implementations of each
function to their respective C files, so now there are no empty C files
anymore.

Same for ldexp and scalbn. Only slight issue that might exist is that
scalbln is a C99 function, and ldexp and scalbn are C89 functions. I'll
ask if that is a problem later.