Replace rep with loops in memcpy. Replace mem clobber with mem output.
The original code had the alignment loop and the remainder loop as
actual loops, not as rep prefixes. Not sure why, but apparently there
was a reason, so the loops are now reproduced.

I also went back and removed the memory clobbers in favor of memory
outputs. This tells the compiler exactly what memory is changed. In the
case of isolated compilation, this should not matter much, since the
memory copied is never read back, so the compiler can't make any smart
decisions about loading it, but in case of LTO, this may be entirely
different.