Add input memory constraints to memcpy.
Of course, movsb not only writes but also reads memory. As before, it
really doesn't matter for isolated compilation (and indeed the compiled
assembler code is exactly the same as before), but it may matter with
LTO, and I at least don't want to stand in the way of that.