Convert memcpy to C.
It turns out that the dummy memory input and output constraints are
actually important. Without them, GCC came to the conclusion that the
second and third __asm__ statements don't matter for anything, because
they only set dst and src, but not *dst and *src, and optimized them
away, since those stores are dead.