Add volatile to asm statement.
In this case, memory is changed that I cannot easily describe with an
lvalue for GCC to accept. So instead, I mark the ASM statement as having
important side effects. It does not mean quite the same thing, but it
prevents it from being elided in case of LTO, if the return value were
unused. Because as it was before, the only job of the ASM statement, as
far as the compiler was concerned, was to set the output variables, and
all of those stores are dead except for the one to dst. But if the
return value were unused, then with LTO, the ASM statement could be
removed.